Renaming StubIndexConfig, stub stuff moving to a separate file
