+++ QuantLib.vcxproj
@@ -541,6 +541,7 @@
     <ClInclude Include="ql\cashflows\rateaveraging.hpp" />
     <ClInclude Include="ql\cashflows\replication.hpp" />
     <ClInclude Include="ql\cashflows\simplecashflow.hpp" />
+    <ClInclude Include="ql\cashflows\stubiborcoupon.hpp" />
     <ClInclude Include="ql\cashflows\timebasket.hpp" />
     <ClInclude Include="ql\cashflows\yoyinflationcoupon.hpp" />
     <ClInclude Include="ql\cashflows\zeroinflationcashflow.hpp" />
@@ -875,7 +876,6 @@
     <ClInclude Include="ql\indexes\swap\jpyliborswap.hpp" />
     <ClInclude Include="ql\indexes\swap\usdliborswap.hpp" />
     <ClInclude Include="ql\indexes\swapindex.hpp" />
-    <ClInclude Include="ql\indexes\weightediborindex.hpp" />
     <ClInclude Include="ql\instruments\all.hpp" />
     <ClInclude Include="ql\instruments\asianoption.hpp" />
     <ClInclude Include="ql\instruments\assetswap.hpp" />
@@ -1994,6 +1994,7 @@
     <ClCompile Include="ql\cashflows\rangeaccrual.cpp" />
     <ClCompile Include="ql\cashflows\replication.cpp" />
     <ClCompile Include="ql\cashflows\simplecashflow.cpp" />
+    <ClCompile Include="ql\cashflows\stubiborcoupon.cpp" />
     <ClCompile Include="ql\cashflows\timebasket.cpp" />
     <ClCompile Include="ql\cashflows\yoyinflationcoupon.cpp" />
     <ClCompile Include="ql\cashflows\zeroinflationcashflow.cpp" />
@@ -2196,7 +2197,6 @@
     <ClCompile Include="ql\indexes\swap\jpyliborswap.cpp" />
     <ClCompile Include="ql\indexes\swap\usdliborswap.cpp" />
     <ClCompile Include="ql\indexes\swapindex.cpp" />
-    <ClCompile Include="ql\indexes\weightediborindex.cpp" />
     <ClCompile Include="ql\instruments\asianoption.cpp" />
     <ClCompile Include="ql\instruments\assetswap.cpp" />
     <ClCompile Include="ql\instruments\averagetype.cpp" />
