+++ ql/CMakeLists.txt
@@ -31,6 +31,7 @@ set(QL_SOURCES
     cashflows/rangeaccrual.cpp
     cashflows/replication.cpp
     cashflows/simplecashflow.cpp
+    cashflows/stubiborcoupon.cpp
     cashflows/timebasket.cpp
     cashflows/yoyinflationcoupon.cpp
     cashflows/zeroinflationcashflow.cpp
@@ -240,7 +241,6 @@ set(QL_SOURCES
     indexes/swap/jpyliborswap.cpp
     indexes/swap/usdliborswap.cpp
     indexes/swapindex.cpp
-    indexes/weightediborindex.cpp
     instrument.cpp
     instruments/asianoption.cpp
     instruments/assetswap.cpp
@@ -1020,6 +1020,7 @@ set(QL_HEADERS
     cashflows/rateaveraging.hpp
     cashflows/replication.hpp
     cashflows/simplecashflow.hpp
+    cashflows/stubiborcoupon.hpp
     cashflows/timebasket.hpp
     cashflows/yoyinflationcoupon.hpp
     cashflows/zeroinflationcashflow.hpp
@@ -1338,7 +1339,6 @@ set(QL_HEADERS
     indexes/swap/jpyliborswap.hpp
     indexes/swap/usdliborswap.hpp
     indexes/swapindex.hpp
-    indexes/weightediborindex.hpp
     instrument.hpp
     instruments/asianoption.hpp
     instruments/assetswap.hpp
