+++ ql/cashflows/all.hpp
@@ -33,6 +33,7 @@
 #include <ql/cashflows/rateaveraging.hpp>
 #include <ql/cashflows/replication.hpp>
 #include <ql/cashflows/simplecashflow.hpp>
+#include <ql/cashflows/stubiborcoupon.hpp>
 #include <ql/cashflows/timebasket.hpp>
 #include <ql/cashflows/yoyinflationcoupon.hpp>
 #include <ql/cashflows/zeroinflationcashflow.hpp>
