+++ ql/cashflows/iborcoupon.cpp
@@ -26,12 +26,10 @@
 #include <ql/cashflows/cashflowvectors.hpp>
 #include <ql/cashflows/couponpricer.hpp>
 #include <ql/cashflows/iborcoupon.hpp>
+#include <ql/cashflows/stubiborcoupon.hpp>
 #include <ql/indexes/interestrateindex.hpp>
 #include <ql/termstructures/yieldtermstructure.hpp>
 #include <ql/optional.hpp>
-#include <algorithm>
-#include <cmath>
-#include <iterator>
 #include <utility>
 
 namespace QuantLib {
@@ -142,172 +140,6 @@ namespace QuantLib {
             v1->visit(*this);
         else
             FloatingRateCoupon::accept(v);
-    }
-
-
-    StubIndexConfig::StubIndexConfig(StubIndexConvention convention,
-                                     std::vector<ext::shared_ptr<IborIndex> > indices)
-    : convention_(convention), indices_(std::move(indices)) {
-        QL_REQUIRE(!indices_.empty(), "no candidate indices provided for stub index selection");
-        for (const auto& candidate : indices_)
-            QL_REQUIRE(candidate, "null candidate index for stub index selection");
-    }
-
-    std::vector<WeightedIborIndex::Component>
-    StubIndexConfig::components(const Date& fixingDate,
-                                const Date& accrualStartDate,
-                                const Date& accrualEndDate) const {
-        QL_REQUIRE(!empty(), "no candidate indices provided for stub index selection");
-
-        // (maturity, index) pairs, sorted by maturity so selection reduces to
-        // a nearest-neighbour or bracket search.
-        std::vector<std::pair<Date, ext::shared_ptr<IborIndex> > > candidates;
-        candidates.reserve(indices_.size());
-        for (const auto& candidate : indices_) {
-            const Date candidateValueDate = candidate->valueDate(fixingDate);
-            QL_REQUIRE(candidateValueDate == accrualStartDate,
-                       candidate->name() << " value date " << candidateValueDate
-                                         << " does not match accrual start date "
-                                         << accrualStartDate);
-            candidates.emplace_back(candidate->maturityDate(candidateValueDate), candidate);
-        }
-        std::sort(candidates.begin(), candidates.end(),
-                  [](const auto& a, const auto& b) { return a.first < b.first; });
-        for (Size i = 1; i < candidates.size(); ++i)
-            QL_REQUIRE(candidates[i - 1].first < candidates[i].first,
-                       "duplicate candidate index maturity " << candidates[i].first);
-
-        if (convention_ == StubIndexConvention::ClosestIndex) {
-            // Since candidates are sorted, ties select the smaller maturity.
-            const auto closest = std::min_element(
-                candidates.begin(), candidates.end(),
-                [accrualEndDate](const auto& a, const auto& b) {
-                    return std::abs(a.first - accrualEndDate) <
-                           std::abs(b.first - accrualEndDate);
-                });
-            return {{closest->second, 1.0}};
-        }
-
-        QL_REQUIRE(convention_ == StubIndexConvention::Interpolated,
-                   "unknown stub index convention");
-
-        const auto longer = std::lower_bound(
-            candidates.begin(), candidates.end(), accrualEndDate,
-            [](const auto& candidate, const Date& d) { return candidate.first < d; });
-        QL_REQUIRE(longer != candidates.end(),
-                   "index maturities do not bracket irregular coupon end date "
-                       << accrualEndDate);
-        if (longer->first == accrualEndDate)
-            return {{longer->second, 1.0}};
-
-        QL_REQUIRE(longer != candidates.begin(),
-                   "index maturities do not bracket irregular coupon end date "
-                       << accrualEndDate);
-        const auto shorter = std::prev(longer);
-        const Real weight = Real(accrualEndDate - shorter->first) /
-                            Real(longer->first - shorter->first);
-        return {{shorter->second, 1.0 - weight}, {longer->second, weight}};
-    }
-
-    ext::shared_ptr<WeightedIborIndex>
-    StubIndexConfig::makeIndex(const Date& fixingDate,
-                               const Date& accrualStartDate,
-                               const Date& accrualEndDate) const {
-        return ext::make_shared<WeightedIborIndex>(
-            components(fixingDate, accrualStartDate, accrualEndDate));
-    }
-
-    static ext::shared_ptr<WeightedIborIndex>
-    makeStubIndex(const StubIndexConfig& config,
-                  const Date& startDate,
-                  const Date& endDate,
-                  Natural fixingDays,
-                  bool isInArrears,
-                  BusinessDayConvention fixingConvention) {
-        QL_REQUIRE(!config.empty(),
-                   "StubIborCoupon requires a non-empty stub index configuration");
-        QL_REQUIRE(!isInArrears,
-                   "stub index selection is not supported for in-arrears coupons");
-
-        const Date fixingDate = config.indices().front()->fixingCalendar().advance(
-            startDate, -static_cast<Integer>(fixingDays), Days, fixingConvention);
-
-        return config.makeIndex(fixingDate, startDate, endDate);
-    }
-
-
-    StubIborCoupon::StubIborCoupon(const Date& paymentDate,
-                                   Real nominal,
-                                   const Date& startDate,
-                                   const Date& endDate,
-                                   Natural fixingDays,
-                                   StubIndexConfig stubIndexConfig,
-                                   Real gearing,
-                                   Spread spread,
-                                   const Date& refPeriodStart,
-                                   const Date& refPeriodEnd,
-                                   const DayCounter& dayCounter,
-                                   bool isInArrears,
-                                   const Date& exCouponDate,
-                                   BusinessDayConvention fixingConvention)
-    : StubIborCoupon(paymentDate, nominal, startDate, endDate, fixingDays,
-                     makeStubIndex(stubIndexConfig, startDate, endDate, fixingDays, isInArrears,
-                                   fixingConvention),
-                     stubIndexConfig, gearing, spread, refPeriodStart, refPeriodEnd, dayCounter,
-                     isInArrears, exCouponDate, fixingConvention) {}
-
-    StubIborCoupon::StubIborCoupon(const Date& paymentDate,
-                                   Real nominal,
-                                   const Date& startDate,
-                                   const Date& endDate,
-                                   Natural fixingDays,
-                                   ext::shared_ptr<WeightedIborIndex> weightedIndex,
-                                   StubIndexConfig stubIndexConfig,
-                                   Real gearing,
-                                   Spread spread,
-                                   const Date& refPeriodStart,
-                                   const Date& refPeriodEnd,
-                                   const DayCounter& dayCounter,
-                                   bool isInArrears,
-                                   const Date& exCouponDate,
-                                   BusinessDayConvention fixingConvention)
-    : IborCoupon(paymentDate, nominal, startDate, endDate, fixingDays, weightedIndex, gearing,
-                 spread, refPeriodStart, refPeriodEnd, dayCounter, isInArrears, exCouponDate,
-                 fixingConvention),
-      stubIndexConfig_(std::move(stubIndexConfig)),
-      weightedIndex_(std::move(weightedIndex)) {
-        for (const auto& component : weightedIndex_->components())
-            QL_REQUIRE(component.first->valueDate(fixingDate()) == accrualStartDate(),
-                       component.first->name()
-                           << " value date for the coupon fixing date " << fixingDate()
-                           << " does not match accrual start date " << accrualStartDate());
-    }
-
-    Rate StubIborCoupon::indexFixing() const {
-        return weightedIndex_->fixing(fixingDate());
-    }
-
-    bool StubIborCoupon::hasFixed() const {
-        const Date today = QuantLib::Settings::instance().evaluationDate();
-        if (fixingDate() > today)
-            return false;
-        if (fixingDate() < today)
-            return true;
-        if (QuantLib::Settings::instance().enforcesTodaysHistoricFixings())
-            return true;
-        const auto& components = weightedIndex_->components();
-        return std::all_of(components.begin(), components.end(),
-                           [this](const auto& component) {
-                               return component.first->hasHistoricalFixing(fixingDate());
-                           });
-    }
-
-    void StubIborCoupon::accept(AcyclicVisitor& v) {
-        auto* v1 = dynamic_cast<Visitor<StubIborCoupon>*>(&v);
-        if (v1 != nullptr)
-            v1->visit(*this);
-        else
-            IborCoupon::accept(v);
     }
 
 
@@ -444,8 +276,8 @@ namespace QuantLib {
         return *this;
     }
 
-    IborLeg& IborLeg::withStubIndexConfig(const StubIndexConfig& config) {
-        stubIndexConfig_ = config;
+    IborLeg& IborLeg::withStubIndexSelection(const StubIndexSelection& selection) {
+        stubIndexSelection_ = ext::make_shared<StubIndexSelection>(selection);
         return *this;
     }
 
@@ -458,7 +290,7 @@ namespace QuantLib {
 			             exCouponPeriod_, exCouponCalendar_, exCouponAdjustment_, exCouponEndOfMonth_,
 			             fixingConvention_);
 
-        if (!stubIndexConfig_.empty()) {
+        if (stubIndexSelection_ && !stubIndexSelection_->empty()) {
             QL_REQUIRE(caps_.empty() && floors_.empty(),
                        "stub index conventions are not supported for capped/floored Ibor legs");
             QL_REQUIRE(!inArrears_,
@@ -480,7 +312,7 @@ namespace QuantLib {
 
                 leg[i] = ext::make_shared<StubIborCoupon>(
                     coupon->date(), coupon->nominal(), coupon->accrualStartDate(),
-                    coupon->accrualEndDate(), coupon->fixingDays(), stubIndexConfig_,
+                    coupon->accrualEndDate(), coupon->fixingDays(), *stubIndexSelection_,
                     coupon->gearing(), coupon->spread(), coupon->referencePeriodStart(),
                     coupon->referencePeriodEnd(), coupon->dayCounter(), coupon->isInArrears(),
                     coupon->exCouponDate(), coupon->fixingConvention());
