+++ ql/experimental/termstructures/crosscurrencyratehelpers.cpp
@@ -98,7 +98,7 @@ namespace QuantLib {
                              const ext::shared_ptr<IborIndex>& idx,
                              Integer paymentLag,
                              std::optional<bool> useIndexedCoupons,
-                             const StubIndexConfig& stubIndexConfig = {}) {
+                             const StubIndexSelection& stubIndexSelection = {}) {
             if (auto overnightIndex = ext::dynamic_pointer_cast<OvernightIndex>(idx)) {
                 return OvernightLeg(schedule, overnightIndex)
                     .withNotionals(1.0)
@@ -108,7 +108,7 @@ namespace QuantLib {
                 .withNotionals(1.0)
                 .withPaymentLag(paymentLag)
                 .withIndexedCoupons(useIndexedCoupons)
-                .withStubIndexConfig(stubIndexConfig);
+                .withStubIndexSelection(stubIndexSelection);
         }
 
         std::pair<Real, Real>
@@ -203,8 +203,8 @@ namespace QuantLib {
         std::optional<Frequency> quoteCurrencyPaymentFrequency,
         std::optional<bool> useIndexedCoupons,
         bool paymentLagOnNotionalExchanges,
-        StubIndexConfig baseStubIndexConfig,
-        StubIndexConfig quoteStubIndexConfig)
+        StubIndexSelection baseStubIndexSelection,
+        StubIndexSelection quoteStubIndexSelection)
     : CrossCurrencySwapRateHelperBase(basis, tenor, fixingDays, std::move(calendar), convention, endOfMonth,
                                       std::move(collateralCurve), paymentLag,
                                       paymentLagOnNotionalExchanges),
@@ -214,13 +214,13 @@ namespace QuantLib {
       paymentFrequency_(normalizedPaymentFrequency(paymentFrequency)),
       quoteCcyPaymentFrequency_(normalizedPaymentFrequency(quoteCurrencyPaymentFrequency)),
       useIndexedCoupons_(useIndexedCoupons),
-      baseStubIndexConfig_(std::move(baseStubIndexConfig)),
-      quoteStubIndexConfig_(std::move(quoteStubIndexConfig)) {
+      baseStubIndexSelection_(std::move(baseStubIndexSelection)),
+      quoteStubIndexSelection_(std::move(quoteStubIndexSelection)) {
         registerWith(baseCcyIdx_);
         registerWith(quoteCcyIdx_);
-        for (const auto& candidate : baseStubIndexConfig_.indices())
+        for (const auto& candidate : baseStubIndexSelection_.indices())
             registerWith(candidate);
-        for (const auto& candidate : quoteStubIndexConfig_.indices())
+        for (const auto& candidate : quoteStubIndexSelection_.indices())
             registerWith(candidate);
 
         CrossCurrencyBasisSwapRateHelperBase::initializeDates();
@@ -232,7 +232,7 @@ namespace QuantLib {
                                                paymentFrequency_);
         baseCcyIborLeg_ = buildFloatingLeg(
             baseCcySchedule_, baseCcyIdx_, paymentLag_, useIndexedCoupons_,
-            baseStubIndexConfig_);
+            baseStubIndexSelection_);
 
         // If no quote-currency payment frequency was given, fall back to the
         // base-currency payment frequency (which may itself be unset, in which
@@ -244,7 +244,7 @@ namespace QuantLib {
                                                 effectiveQuoteCcyFreq);
         quoteCcyIborLeg_ = buildFloatingLeg(
             quoteCcySchedule_, quoteCcyIdx_, paymentLag_, useIndexedCoupons_,
-            quoteStubIndexConfig_);
+            quoteStubIndexSelection_);
 
         initializeDatesFromLegs(baseCcyIborLeg_, quoteCcyIborLeg_);
     }
@@ -276,8 +276,8 @@ namespace QuantLib {
         std::optional<Frequency> quoteCurrencyPaymentFrequency,
         std::optional<bool> useIndexedCoupons,
         bool paymentLagOnNotionalExchanges,
-        StubIndexConfig baseStubIndexConfig,
-        StubIndexConfig quoteStubIndexConfig)
+        StubIndexSelection baseStubIndexSelection,
+        StubIndexSelection quoteStubIndexSelection)
     : CrossCurrencyBasisSwapRateHelperBase(basis,
                                            tenor,
                                            fixingDays,
@@ -294,8 +294,8 @@ namespace QuantLib {
                                            quoteCurrencyPaymentFrequency,
                                            useIndexedCoupons,
                                            paymentLagOnNotionalExchanges,
-                                           std::move(baseStubIndexConfig),
-                                           std::move(quoteStubIndexConfig)) {
+                                           std::move(baseStubIndexSelection),
+                                           std::move(quoteStubIndexSelection)) {
         buildSwap();
     }
 
@@ -314,8 +314,8 @@ namespace QuantLib {
             paymentLag_, paymentLag_, false, Null<Natural>(), false, 0,
             RateAveraging::Compound, false, Null<Natural>(), false, 0,
             RateAveraging::Compound, false, useIndexedCoupons_,
-            paymentLagOnNotionalExchanges_, baseStubIndexConfig_,
-            quoteStubIndexConfig_);
+            paymentLagOnNotionalExchanges_, baseStubIndexSelection_,
+            quoteStubIndexSelection_);
         swap_->setPricingEngine(ext::make_shared<DiscountingConstNotionalCrossCurrencySwapEngine>(
             quoteCcyIdx_->currency(), quoteCcyLegDiscountHandle(),
             baseCcyIdx_->currency(), baseCcyLegDiscountHandle(),
@@ -364,8 +364,8 @@ namespace QuantLib {
         Natural fxResetFixingDays,
         Calendar fxResetFixingCalendar,
         std::optional<bool> useIndexedCoupons,
-        StubIndexConfig baseStubIndexConfig,
-        StubIndexConfig quoteStubIndexConfig)
+        StubIndexSelection baseStubIndexSelection,
+        StubIndexSelection quoteStubIndexSelection)
     : CrossCurrencyBasisSwapRateHelperBase(basis,
                                            tenor,
                                            fixingDays,
@@ -382,8 +382,8 @@ namespace QuantLib {
                                            quoteCurrencyPaymentFrequency,
                                            useIndexedCoupons,
                                            false,
-                                           std::move(baseStubIndexConfig),
-                                           std::move(quoteStubIndexConfig)),
+                                           std::move(baseStubIndexSelection),
+                                           std::move(quoteStubIndexSelection)),
       isFxBaseCurrencyLegResettable_(isFxBaseCurrencyLegResettable),
       fxResetFixingDays_(fxResetFixingDays), fxResetFixingCalendar_(std::move(fxResetFixingCalendar)) {
         buildSwap();
@@ -407,7 +407,7 @@ namespace QuantLib {
             convention_, convention_, false, Null<Natural>(), false, 0,
             RateAveraging::Compound, false, Null<Natural>(), false, 0,
             RateAveraging::Compound, false, useIndexedCoupons_,
-            baseStubIndexConfig_, quoteStubIndexConfig_);
+            baseStubIndexSelection_, quoteStubIndexSelection_);
         swap_->setPricingEngine(ext::make_shared<DiscountingMtMCrossCurrencyBasisSwapEngine>(
             quoteCcyIdx_->currency(), quoteCcyLegDiscountHandle(),
             baseCcyIdx_->currency(), baseCcyLegDiscountHandle(),
@@ -448,7 +448,7 @@ namespace QuantLib {
         Integer paymentLag,
         std::optional<bool> useIndexedCoupons,
         std::optional<Frequency> floatPaymentFrequency,
-        StubIndexConfig floatStubIndexConfig)
+        StubIndexSelection floatStubIndexSelection)
     : CrossCurrencySwapRateHelperBase(fixedRate, tenor, fixingDays, calendar, convention, endOfMonth,
                                       collateralCurve, paymentLag),
       fixedFrequency_(fixedFrequency),
@@ -457,11 +457,11 @@ namespace QuantLib {
       collateralOnFixedLeg_(collateralOnFixedLeg),
       useIndexedCoupons_(useIndexedCoupons),
       floatPaymentFrequency_(normalizedPaymentFrequency(floatPaymentFrequency)),
-      floatStubIndexConfig_(std::move(floatStubIndexConfig)) {
+      floatStubIndexSelection_(std::move(floatStubIndexSelection)) {
 
         QL_REQUIRE(floatIndex_, "floating index required");
         registerWith(floatIndex_);
-        for (const auto& candidate : floatStubIndexConfig_.indices())
+        for (const auto& candidate : floatStubIndexSelection_.indices())
             registerWith(candidate);
 
         initializeDates();
@@ -501,7 +501,7 @@ namespace QuantLib {
             calendar_,
             false, false, Null<Natural>(), false, 0,
             RateAveraging::Compound, useIndexedCoupons_,
-            floatStubIndexConfig_
+            floatStubIndexSelection_
         );
         auto engine = ext::make_shared<DiscountingConstNotionalCrossCurrencySwapEngine>(
             floatIndex_->currency(), floatingLegDiscountHandle(),
