+++ ql/experimental/termstructures/crosscurrencyratehelpers.hpp
@@ -26,6 +26,7 @@
 #ifndef quantlib_crosscurrencyratehelpers_hpp
 #define quantlib_crosscurrencyratehelpers_hpp
 
+#include <ql/cashflows/stubiborcoupon.hpp>
 #include <ql/termstructures/yield/ratehelpers.hpp>
 #include <ql/instruments/constnotionalcrosscurrencybasisswap.hpp>
 #include <ql/instruments/constnotionalcrosscurrencyfixedvsfloatingswap.hpp>
@@ -87,8 +88,8 @@ namespace QuantLib {
                                              std::optional<Frequency> quoteCurrencyPaymentFrequency = std::nullopt,
                                              std::optional<bool> useIndexedCoupons = std::nullopt,
                                              bool paymentLagOnNotionalExchanges = false,
-                                             StubIndexConfig baseStubIndexConfig = {},
-                                             StubIndexConfig quoteStubIndexConfig = {});
+                                             StubIndexSelection baseStubIndexSelection = {},
+                                             StubIndexSelection quoteStubIndexSelection = {});
 
         void initializeDates() override;
         const Handle<YieldTermStructure>& baseCcyLegDiscountHandle() const;
@@ -101,8 +102,8 @@ namespace QuantLib {
         std::optional<Frequency> paymentFrequency_;
         std::optional<Frequency> quoteCcyPaymentFrequency_;
         std::optional<bool> useIndexedCoupons_;
-        StubIndexConfig baseStubIndexConfig_;
-        StubIndexConfig quoteStubIndexConfig_;
+        StubIndexSelection baseStubIndexSelection_;
+        StubIndexSelection quoteStubIndexSelection_;
 
         Schedule baseCcySchedule_;
         Schedule quoteCcySchedule_;
@@ -154,13 +155,13 @@ namespace QuantLib {
                 the coupons: the final exchange settles together with the final
                 coupon and the initial exchange falls on the lagged settlement
                 date (default: false).
-            \param baseStubIndexConfig
+            \param baseStubIndexSelection
                 index selection applied to irregular coupons of the base-currency
-                leg when it is an Ibor leg (see StubIndexConfig).  The candidate
+                leg when it is an Ibor leg (see StubIndexSelection).  The candidate
                 indices must use exogenous forwarding curves; the default prices
                 broken periods off the leg's own index.
-            \param quoteStubIndexConfig
-                as baseStubIndexConfig, for the quote-currency leg.
+            \param quoteStubIndexSelection
+                as baseStubIndexSelection, for the quote-currency leg.
             In both frequency parameters, \c NoFrequency is accepted as a synonym for
             an unset (null) value.
         */
@@ -181,8 +182,8 @@ namespace QuantLib {
             std::optional<Frequency> quoteCurrencyPaymentFrequency = std::nullopt,
             std::optional<bool> useIndexedCoupons = std::nullopt,
             bool paymentLagOnNotionalExchanges = false,
-            StubIndexConfig baseStubIndexConfig = {},
-            StubIndexConfig quoteStubIndexConfig = {});
+            StubIndexSelection baseStubIndexSelection = {},
+            StubIndexSelection quoteStubIndexSelection = {});
         //! \name RateHelper interface
         //@{
         Real impliedQuote() const override;
@@ -235,13 +236,13 @@ namespace QuantLib {
                 \p calendar is used if this is empty.
             \param useIndexedCoupons
                 if provided, overrides the global IborCoupon setting for both legs.
-            \param baseStubIndexConfig
+            \param baseStubIndexSelection
                 index selection applied to irregular coupons of the base-currency
-                leg when it is an Ibor leg (see StubIndexConfig).  The candidate
+                leg when it is an Ibor leg (see StubIndexSelection).  The candidate
                 indices must use exogenous forwarding curves; the default prices
                 broken periods off the leg's own index.
-            \param quoteStubIndexConfig
-                as baseStubIndexConfig, for the quote-currency leg.
+            \param quoteStubIndexSelection
+                as baseStubIndexSelection, for the quote-currency leg.
             In both frequency parameters, \c NoFrequency is accepted as a synonym for
             an unset (null) value.
         */
@@ -263,8 +264,8 @@ namespace QuantLib {
                                             Natural fxResetFixingDays = 0,
                                             Calendar fxResetFixingCalendar = Calendar(),
                                             std::optional<bool> useIndexedCoupons = std::nullopt,
-                                            StubIndexConfig baseStubIndexConfig = {},
-                                            StubIndexConfig quoteStubIndexConfig = {});
+                                            StubIndexSelection baseStubIndexSelection = {},
+                                            StubIndexSelection quoteStubIndexSelection = {});
         //! \name RateHelper interface
         //@{
         Real impliedQuote() const override;
@@ -319,8 +320,8 @@ namespace QuantLib {
     has no payment frequency of its own, so one must be given explicitly.
     \c NoFrequency is accepted as a synonym for an unset (null) value.
 
-    The floatStubIndexConfig parameter selects the indices used for irregular
-    coupons of the floating leg when it is an Ibor leg (see StubIndexConfig).
+    The floatStubIndexSelection parameter selects the indices used for irregular
+    coupons of the floating leg when it is an Ibor leg (see StubIndexSelection).
     */
     class ConstNotionalCrossCurrencySwapRateHelper : public CrossCurrencySwapRateHelperBase {
       public:
@@ -339,7 +340,7 @@ namespace QuantLib {
             Integer paymentLag = 0,
             std::optional<bool> useIndexedCoupons = std::nullopt,
             std::optional<Frequency> floatPaymentFrequency = std::nullopt,
-            StubIndexConfig floatStubIndexConfig = {});
+            StubIndexSelection floatStubIndexSelection = {});
 
         Real impliedQuote() const override;
         void accept(AcyclicVisitor&) override;
@@ -359,7 +360,7 @@ namespace QuantLib {
         bool collateralOnFixedLeg_;
         std::optional<bool> useIndexedCoupons_;
         std::optional<Frequency> floatPaymentFrequency_;
-        StubIndexConfig floatStubIndexConfig_;
+        StubIndexSelection floatStubIndexSelection_;
 
         ext::shared_ptr<ConstNotionalCrossCurrencyFixedVsFloatingSwap> xccySwap_;
     };
