+++ ql/indexes/all.hpp
@@ -9,7 +9,6 @@
 #include <ql/indexes/interestrateindex.hpp>
 #include <ql/indexes/region.hpp>
 #include <ql/indexes/swapindex.hpp>
-#include <ql/indexes/weightediborindex.hpp>
 
 #include <ql/indexes/ibor/all.hpp>
 #include <ql/indexes/inflation/all.hpp>
