+++ ql/instruments/constnotionalcrosscurrencybasisswap.cpp
@@ -41,16 +41,16 @@ ConstNotionalCrossCurrencyBasisSwap::ConstNotionalCrossCurrencyBasisSwap(
                                      const bool telescopicValueDates,
                                      std::optional<bool> useIndexedCoupons,
                                      const bool paymentLagOnNotionalExchanges,
-                                     StubIndexConfig payStubIndexConfig,
-                                     StubIndexConfig recStubIndexConfig)
+                                     StubIndexSelection payStubIndexSelection,
+                                     StubIndexSelection recStubIndexSelection)
     : ConstNotionalCrossCurrencySwap(2), payNominal_(payNominal), payCurrency_(std::move(payCurrency)), paySchedule_(std::move(paySchedule)),
       payIndex_(payIndex), paySpread_(paySpread), payGearing_(payGearing), recNominal_(recNominal),
       recCurrency_(std::move(recCurrency)), recSchedule_(std::move(recSchedule)), recIndex_(recIndex), recSpread_(recSpread),
       recGearing_(recGearing), payPaymentLag_(payPaymentLag), recPaymentLag_(recPaymentLag),
       useIndexedCoupons_(useIndexedCoupons),
       paymentLagOnNotionalExchanges_(paymentLagOnNotionalExchanges),
-      payStubIndexConfig_(std::move(payStubIndexConfig)),
-      recStubIndexConfig_(std::move(recStubIndexConfig)),
+      payStubIndexSelection_(std::move(payStubIndexSelection)),
+      recStubIndexSelection_(std::move(recStubIndexSelection)),
       payCompoundSpread_(payCompoundSpread), payLookbackDays_(payLookbackDays),
       payObservationShift_(payObservationShift), payLockoutDays_(payLockoutDays),
       payAveragingMethod_(payAveragingMethod), recCompoundSpread_(recCompoundSpread),
@@ -84,7 +84,7 @@ void ConstNotionalCrossCurrencyBasisSwap::initialize() {
                        .withGearings(payGearing_)
                        .withPaymentLag(payPaymentLag_)
                        .withIndexedCoupons(useIndexedCoupons_)
-                       .withStubIndexConfig(payStubIndexConfig_);
+                       .withStubIndexSelection(payStubIndexSelection_);
     }
     payer_[0] = -1.0;
     currencies_[0] = payCurrency_;
@@ -111,7 +111,7 @@ void ConstNotionalCrossCurrencyBasisSwap::initialize() {
                        .withGearings(recGearing_)
                        .withPaymentLag(recPaymentLag_)
                        .withIndexedCoupons(useIndexedCoupons_)
-                       .withStubIndexConfig(recStubIndexConfig_);
+                       .withStubIndexSelection(recStubIndexSelection_);
     }
     payer_[1] = +1.0;
     currencies_[1] = recCurrency_;
